# arm/rp2040: SPI driver DMA transfer support