## Summary
Support DMA transfer for RP2040 SPI driver.

## Impact
rp2040 only

## Testing
Tested with `raspberrypi-pico:spisd` configuration.
Read/Write access OK.