@acassis 2021-03-08
Hi @yunkya2 about "Tested with raspberrypi-pico:spisd configuration" how much it improved the speed? Did you do some measurements? BR, Alan