@yunkya2 2021-03-08
@acassis I have tested by writing/reading 10MB file.

Without DMA: (`CONFIG_RP2040_SPI_DMA is not set`)
```
nsh> rm test
nsh> time "dd if=/dev/zero of=test bs=32768 count=320"

6.5500 sec

nsh> time "dd if=test of=/dev/null bs=32768 count=320"

5.6200 sec
```
- Write: 12.8Mbps
- Read: 14.9Mbps

 
With DMA: (`CONFIG_RP2040_SPI_DMA=y`)
```
nsh> rm test
nsh> time "dd if=/dev/zero of=test bs=32768 count=320"

6.1800 sec

nsh> time "dd if=test of=/dev/null bs=32768 count=320"

5.2500 sec
```

- Write: 13.6Mbps
- Read: 16.0Mbps

The performance improvement is not so high.
Rather than it, the major improvement is the decrease of CPU load while transfer.