@acassis 2021-03-09
I see! Thank you @yunkya2 ! Any improvement is a good progress! Ye the decrease of the CPU load is the best result!